requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
gunicorn==24.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
numpy==2.1.3
packaging==26.0
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.17.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.3
Werkzeug==3.1.5
mysql-connector-python==9.0.0